First working phaser & flanger
Phaser and flanger both process sound without major artifacts or noise

+++ fxobjects.cpp
@@ -1006,6 +1006,7 @@ void AudioDetector::setReleaseTime(double release_in_ms, bool forceCalc)
 /**
 \brief generates the oscillator output for one sample interval; note that there are multiple outputs.
 */
+// changed to work with float
 const SignalGenData LFO::renderAudioOutput()
 {
 	// --- always first!
@@ -1024,13 +1025,13 @@ const SignalGenData LFO::renderAudioOutput()
 	if (waveform == generatorWaveform::kSin)
 	{
 		// --- calculate normal angle
-		double angle = modCounter*2.0*kPi - kPi;
+		float angle = modCounter*2.0f*kPi - kPi;
 
 		// --- norm output with parabolicSine approximation
 		output.normalOutput = parabolicSine(-angle);
 
 		// --- calculate QP angle
-		angle = modCounterQP*2.0*kPi - kPi;
+		angle = modCounterQP*2.0f*kPi - kPi;
 
 		// --- calc QP output
 		output.quadPhaseOutput_pos = parabolicSine(-angle);
@@ -1041,13 +1042,13 @@ const SignalGenData LFO::renderAudioOutput()
 		output.normalOutput = unipolarToBipolar(modCounter);
 
 		// bipolar triagle
-		output.normalOutput = 2.0*fabs(output.normalOutput) - 1.0;
+		output.normalOutput = 2.0f*(float)fabs(output.normalOutput) - 1.0f;
 
 		// -- quad phase
 		output.quadPhaseOutput_pos = unipolarToBipolar(modCounterQP);
 
 		// bipolar triagle
-		output.quadPhaseOutput_pos = 2.0*fabs(output.quadPhaseOutput_pos) - 1.0;
+		output.quadPhaseOutput_pos = 2.0f*(float)fabs(output.quadPhaseOutput_pos) - 1.0f;
 	}
 	else if (waveform == generatorWaveform::kSaw)
 	{
